Version 2: Cambios en las enumeraciones, en la forma de añadir un personaje al listado, cambios en los menús, funcionalidad de la GUI.

+++ TrabajoFinalMunnozCordonEstela/src/trabajoFinal/personajes/Arquero.java
@@ -67,7 +67,7 @@ public class Arquero extends Personaje implements Razable {
 	 * @throws RazaNoValidaException Error por raza incorrecta.
 	 */
 	protected Arquero(String nombre, String apellido, Sexo sexo,
-			Zodiaco zodiaco, int edad, float altura, float peso, Region region, Raza raza, String descripcion)
+			Zodiaco zodiaco, int edad, double altura, double peso, Region region, Raza raza, String descripcion)
 			throws ApellidoNoValidoException, NombreNoValidoException,
 			RegionNoValidaException, SexoNoValidoException, AlturaNoValidaException, 
 			PesoNoValidoException, ZodiacoNoValidoException, EdadNoValidaException,
